+++ pipelex/plugins/plugin_sdk_registry.py
@@ -1,25 +1,8 @@
 from typing import Any
 
-from pydantic import BaseModel, Field, RootModel
+from pydantic import Field, RootModel
 
-from pipelex.cogt.model_backends.model_spec import InferenceModelSpec
-
-
-class Plugin(BaseModel):
-    sdk: str
-    backend: str
-
-    @property
-    def sdk_handle(self) -> str:
-        return f"{self.sdk}@{self.backend}"
-
-    @classmethod
-    def make_for_inference_model(cls, inference_model: InferenceModelSpec) -> "Plugin":
-        return Plugin(
-            sdk=inference_model.sdk,
-            backend=inference_model.backend_name,
-        )
-
+from pipelex.plugins.plugin import Plugin
 
 PluginSdkRegistryRoot = dict[str, Any]
