+++ tests/integration/pipelex/conftest.py
@@ -19,7 +19,7 @@ def llm_preset_id(request: pytest.FixtureRequest) -> str:
 @pytest.fixture(
     params=[
         # "gpt-4o",
-        "gpt-4o-mini",
+        # "gpt-4o-mini",
         # "gpt-4-5-preview",
         # "o1",
         # "o1-mini",
@@ -52,7 +52,7 @@ def llm_preset_id(request: pytest.FixtureRequest) -> str:
         # "gemini-2.5-flash",
         # "gemini-2.5-pro",
         # "openai/gpt-oss-120b",
-        # "meta-llama/llama-4-scout-17b-16e-instruct",
+        "meta-llama/llama-4-scout-17b-16e-instruct",
         # "meta-llama/llama-4-maverick-17b-128e-instruct",
         # "moonshotai/kimi-k2-instruct-0905",
     ],
